Fixed incorrect parameters parsing. Add hashCode, equals and toString methods

+++ src/factory/GiftFactory.java
@@ -13,7 +13,7 @@ public class GiftFactory {
 		String params = parser.getParams(requaredGiftElement);
 		switch (objectType) {
 		case "Sweet":
-			long sweethess = Long.valueOf(params);
+			int sweethess = Integer.valueOf(params);
 			return new Sweet(sweethess);
 		case "Wrapping":
 			String[] patternAndColor = params.split("[ \\t]+");
